requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
deep-translator==1.11.4
tqdm==4.66.1
boto3==1.35.94
python-dotenv==1.0.1
